requirements.txt
allure-pytest==2.11.1
allure-python-commons==2.11.1
Appium-Python-Client==2.1.4
asgiref==3.4.1
async-generator==1.10
attrs==21.4.0
blessings==1.7
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
codecov==2.1.12
coloredlogs==15.0.1
colour-runner==0.1.1
coverage==6.5.0
cryptography==36.0.1
deepdiff==5.8.1
distlib==0.3.6
Django==3.2.9
djangorestframework==3.12.4
filelock==3.8.0
glob2==0.7
h11==0.13.0
humanfriendly==10.0
idna==3.3
iniconfig==1.1.1
Mako==1.2.3
MarkupSafe==2.1.1
ordered-set==4.1.0
outcome==1.1.0
packaging==21.3
parse==1.19.0
parse-type==0.6.0
Pillow==8.4.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.1.0
pytest-bdd==6.0.1
pytest-html==3.1.1
pytest-metadata==2.0.2
python-dotenv==0.21.0
pytz==2021.3
requests==2.26.0
rootpath==0.1.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
termcolor==2.0.1
tomli==2.0.1
tox==3.26.0
tqdm==4.64.1
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.7
virtualenv==20.16.5
webdriver-manager==3.8.3
wsproto==1.1.0
